Make step_id be of type uint32_t everywhere (was uint16_t in some places)

+++ branches/blcr/src/common/slurm_protocol_defs.h
@@ -766,7 +766,7 @@ typedef struct slurm_node_registration_status_msg {
 	uint32_t tmp_disk;
 	uint32_t job_count;	/* number of associate job_id's */
 	uint32_t *job_id;	/* IDs of running job (if any) */
-	uint16_t *step_id;	/* IDs of running job steps (if any) */
+	uint32_t *step_id;	/* IDs of running job steps (if any) */
 	uint32_t status;	/* node status code, same as return codes */
 	uint16_t startup;	/* slurmd just restarted */
 	switch_node_info_t switch_nodeinfo;	/* set only if startup != 0 */
